still working on tree checking function
