+++ main.cpp
@@ -19,8 +19,9 @@
 #include <string.h>
 #include <windows.h>//needed for call to "exit()"
 #include <math.h>
-#include "xo_classes.h"//include prototype declarations for class functions.
+#include "xo_game.h"//include prototype declarations for game functions.
                             //the classes themselves are written in other .cpp files
+#include "xo_computer.h"//include prototype declarations for AI classes
 
 using namespace std;
