+++ xo_computer.cpp
@@ -1,9 +1,12 @@
-
-#include "xo_classes.h"
-#include <algorithm> // needed for calls to max()
 #include <stdio.h>
 #include <iostream>
 #include <string.h>
+#include <windows.h>//needed for call to "exit()"
+#include <math.h>
+#include "xo_game.h"//include prototype declarations for game functions.
+                            //the classes themselves are written in other .cpp files
+#include "xo_computer.h"//include prototype declarations for AI classes
+
 using namespace std;
 
 //--------
