+++ xo_game.cpp
@@ -1,9 +1,13 @@
-#include "xo_classes.h"
 #include <stdio.h>
 #include <iostream>
+#include <string.h>
+#include <windows.h>//needed for call to "exit()"
+#include <math.h>
+#include "xo_game.h"//include prototype declarations for game functions.
+                            //the classes themselves are written in other .cpp files
+#include "xo_computer.h"//include prototype declarations for AI classes
 
 using namespace std;
-
 //------------------
 //Game_State functions
 //------------------------
